put the model in separate model folder
